客房类型管理

+++ src/main/java/com/cy/pj/sys/controller/SysRoleController.java
@@ -31,7 +31,7 @@ public class SysRoleController {
 	
 	@RequestMapping("doFindObjectById")
     @ResponseBody
-    public JsonResult doFindObjectById(Integer id) {
+    public JsonResult doFindObjectById(Long id) {
         return new JsonResult(sysRoleService.findObjectById(id));
     }
